+++ enclave/sgx/link.c
@@ -3,6 +3,7 @@
 
 #include <openenclave/enclave.h>
 #include <openenclave/internal/malloc.h>
+#include "../asym_keys.h"
 #include "report.h"
 
 //
@@ -19,6 +20,8 @@ const void* oe_link_enclave(void)
 {
     static const void* symbols[] = {
         oe_handle_verify_report,
+        oe_handle_get_public_key_by_policy,
+        oe_handle_get_public_key,
 #if defined(OE_USE_DEBUG_MALLOC)
         oe_debug_malloc_check,
 #endif /* defined(OE_USE_DEBUG_MALLOC) */